ML/requirements.txt
flask==3.0.2
transformers==4.37.2
torch==2.2.0
numpy==1.24.3
Pillow==10.2.0
nltk==3.8.1 